// astr-ai-code-mother-frontend/lib/api/generated/model/chatHistoryQueryRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface ChatHistoryQueryRequest {
  pageNum?: number;
  pageSize?: number;
  sortField?: string;
  sortOrder?: string;
  id?: number;
  message?: string;
  messageType?: string;
  appId?: number;
  userId?: number;
}